requirements.txt
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
hydra-core==1.3.2
matplotlib==3.7.3
numpy==1.24.4
omegaconf==2.3.0
pytorch_lightning==2.1.2
scikit_learn==1.3.1
timm==0.4.12
tqdm==4.66.1
setproctitle
tensorboard
stable-baselines3
envpool
imageio
mujoco
rl_zoo3
gymnasium[atari]
gymnasium[accept-rom-license]
diffusers
transformers
ipykernel
jupyter
ipython
tensorboard
opencv-python